+++ config/analytics.php
@@ -11,7 +11,7 @@ return [
      * Path to the client secret json file. Take a look at the README of this package
      * to learn how to get this file.
      */
-    'service_account_credentials_json' => storage_path('app/laravel-google-analytics/service-account-credentials.json'),
+    'service_account_credentials_json' => storage_path('app/analytics/service-account-credentials.json'),
 
     /*
      * The amount of minutes the Google API responses will be cached.
@@ -20,8 +20,13 @@ return [
     'cache_lifetime_in_minutes' => 60 * 24,
 
     /*
-     * The directory where the underlying Google_Client will store it's cache files.
+     * Here you may configure the "store" that the underlying Google_Client will
+     * use to store it's data.  You may also add extra parameters that will
+     * be passed on setCacheConfig (see docs for google-api-php-client).
+     *
+     * Optional parameters: "lifetime", "prefix"
      */
-    'cache_location' => storage_path('app/laravel-google-analytics/google-cache/'),
-
+    'cache' => [
+        'store' => 'file',
+    ],
 ];