+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to Laravel-Analytics will be documented in this file
 
+## 3.1.0 - 2017-08-31
+- add compatibility with Laravel 5.5
+
+## 3.0.1 - 2017-06-16
+- make publishing the config file optional
+
+## 3.0.0 - 2016-08-23
+- add support for `v2` of the Google API
+- renamed config file from `laravel-analytics.php` to `analytics.php`
+- dropped support for anything lower than Laravel 5.4
+
 ## 2.4.0 - 2017-01-23
 - add support for Laravel 5.4
 - dropped support for anything lower than Laravel 5.3